Fix all `mypy` issues

+++ videbo/misc.py
@@ -15,11 +15,11 @@ logger = logging.getLogger('videbo-misc')
 MEGA = 1024 * 1024
 
 
-async def get_free_disk_space(path: str) -> int:
+async def get_free_disk_space(path: str) -> float:
     """Get free disk space in the given path. Returns MB."""
     st = await asyncio.get_running_loop().run_in_executor(None, os.statvfs, path)
     free_bytes = st.f_bavail * st.f_frsize
-    return int(free_bytes / MEGA)
+    return free_bytes / MEGA
 
 
 def sanitize_filename(filename: str) -> str:
